Automatically detect if a linux system that uses rpm is running on systemd or sysv (#3739)
Issue: https://github.com/Velocidex/velociraptor/issues/3451

rpm will figure out if the Linux OS is using sysv or systemd when
creating velociraptor client

---------

+++ bin/rpm.go
@@ -34,9 +34,6 @@ var (
 	server_rpm_command_binary = server_rpm_command.Flag(
 		"binary", "The binary to package").String()
 
-	client_rpm_command_use_sysv = client_rpm_command.Flag(
-		"use_sysv", "Use SysV style services (CentOS 6)").Bool()
-
 	client_rpm_command_output = client_rpm_command.Flag(
 		"output", "Filename to output").String()
 
@@ -102,9 +99,9 @@ setcap CAP_SYS_RESOURCE,CAP_NET_BIND_SERVICE=+eip /usr/local/bin/velociraptor
 . /etc/rc.d/init.d/functions
 
 RETVAL=0
-prog="velociraptor"
+prog="velociraptor_client"
 lockfile=/var/lock/subsys/$prog
-VELOCIRAPTOR=/usr/local/bin/velociraptor
+VELOCIRAPTOR=/usr/local/bin/velociraptor_client
 VELOCIRAPTOR_CONFIG=/etc/velociraptor/client.config.yaml
 PID_FILE=/var/run/velociraptor.pid
 
@@ -215,7 +212,8 @@ exit $RETVAL
 	}
 )
 
-// Systemd based start up scripts (Centos 7, 8)
+// Use Systemd based start up scripts (Centos 7, 8) if /bin/systemctl exists on OS
+// otherwise use Simple startup scripts for SysV-style init systems (Centos 6)
 func doClientRPM() error {
 	// Disable logging when creating a package - we may not create the
 	// package on the same system where the logs should go.
@@ -306,29 +304,61 @@ func doClientRPM() error {
 	config_path := "/etc/velociraptor/client.config.yaml"
 	velociraptor_bin := "/usr/local/bin/velociraptor_client"
 
-	r.AddFile(
-		rpmpack.RPMFile{
-			Name: "/etc/systemd/system/velociraptor_client.service",
-			Body: []byte(fmt.Sprintf(
-				client_service_definition, velociraptor_bin, config_path)),
-			Mode:  0644,
-			Owner: "root",
-			Group: "root",
-		})
-
-	r.AddPostin(`/bin/systemctl enable velociraptor_client.service
-/bin/systemctl start velociraptor_client.service
+	r.AddPrein(`
+if ! [ -f /bin/systemctl ] ; then
+   getent group velociraptor >/dev/null || groupadd -g 115 -r velociraptor || :
+   getent passwd velociraptor >/dev/null || \
+   useradd -c "Privilege-separated Velociraptor" -u 115 -g velociraptor  -s /sbin/nologin \
+     -s /sbin/nologin -r -d /var/empty/velociraptor velociraptor 2> /dev/null || :
+fi
 `)
 
-	// check for upgrade vs uninstall
-	r.AddPreun(`
-if [ $1 == 1 ] ; then
-    /bin/systemctl restart velociraptor_client.service
-fi
+	r.AddPostin(fmt.Sprintf(`
+if [ -f /bin/systemctl ] ; then
+cat << SYSTEMDSCRIPT > /etc/systemd/system/velociraptor_client.service
+%s
+SYSTEMDSCRIPT
+/bin/systemctl enable velociraptor_client.service
+/bin/systemctl start velociraptor_client.service
+else
+cat << SYSVSCRIPT > /etc/rc.d/init.d/velociraptor
+%s
+SYSVSCRIPT
+/bin/chmod +x /etc/rc.d/init.d/velociraptor
 
-if [ $1 == 0 ] ; then
-    /bin/systemctl disable velociraptor_client.service
-    /bin/systemctl stop velociraptor_client.service
+## Set it to start at boot
+/sbin/chkconfig --add velociraptor
+
+## Start the service immediately
+service velociraptor start
+fi
+`, fmt.Sprintf(client_service_definition, velociraptor_bin, config_path),
+		escape_sh(rpm_sysv_client_service_definition)))
+
+	r.AddPreun(`
+if [ -f /bin/systemctl ]; then
+    if [ $1 == 1 ] ; then
+        /bin/systemctl restart velociraptor_client.service
+    fi
+
+    if [ $1 == 0 ] ; then
+        /bin/systemctl disable velociraptor_client.service
+        /bin/systemctl stop velociraptor_client.service
+    fi
+else
+    if [ $1 = 0 ] ; then
+        /sbin/service velociraptor stop > /dev/null 2>&1 || :
+        /sbin/chkconfig --del velociraptor
+    fi
+fi
+`)
+
+	r.AddPostun(`
+if [ -f /bin/systemctl ] ; then
+	rm /etc/systemd/system/velociraptor_client.service
+else
+	/sbin/service velociraptor start  > /dev/null 2>&1 || :
+	rm /etc/rc.d/init.d/velociraptor
 fi
 `)
 
@@ -492,138 +522,11 @@ func doSingleServerRPM(
 	return r.Write(fd)
 }
 
-// Simple startup scripts for SysV-style init systems (Centos 6)
-func doClientSysVRPM() error {
-	// Disable logging when creating a package - we may not create the
-	// package on the same system where the logs should go.
-	logging.DisableLogging()
-
-	config_obj, err := makeDefaultConfigLoader().
-		WithRequiredClient().LoadAndValidate()
-	if err != nil {
-		return fmt.Errorf("Unable to load config file: %w", err)
-	}
-
-	ctx, cancel := install_sig_handler()
-	defer cancel()
-
-	sm, err := startup.StartToolServices(ctx, config_obj)
-	defer sm.Close()
-
-	if err != nil {
-		return err
-	}
-
-	config_file_yaml, err := yaml.Marshal(getClientConfig(config_obj))
-	if err != nil {
-		return err
-	}
-
-	input := *client_rpm_command_binary
-
-	if input == "" {
-		input, err = os.Executable()
-		if err != nil {
-			return fmt.Errorf("Unable to find executable: %w", err)
-		}
-	}
-
-	e, err := elf.Open(input)
-	if err != nil {
-		return fmt.Errorf("Unable to parse ELF executable: %w", err)
-	}
-
-	arch, ok := rpmArchMap[e.Machine.String()]
-	if !ok {
-		return fmt.Errorf("unknown binary architecture: %q", e.Machine.String())
-	}
-
-	binary_content, err := os.ReadFile(input)
-	if err != nil {
-		return fmt.Errorf("Unable to read executable: %w", err)
-	}
-
-	version := strings.ReplaceAll(constants.VERSION, "-", ".")
-
-	output_path := fmt.Sprintf("velociraptor_client_%s_%s.rpm", version, arch)
-	if *client_rpm_command_output != "" {
-		output_path = *client_rpm_command_output
-	}
-
-	fmt.Printf("Creating SysV-init client package at %s\n", output_path)
-
-	r, err := rpmpack.NewRPM(rpmpack.RPMMetaData{
-		Name:    "velociraptor-client",
-		Version: version,
-		Release: *rpm_command_release,
-		Arch:    arch,
-	})
-	if err != nil {
-		return fmt.Errorf("Unable to create RPM: %w", err)
-	}
-	r.AddFile(
-		rpmpack.RPMFile{
-			Name:  "/etc/velociraptor/client.config.yaml",
-			Mode:  0600,
-			Body:  config_file_yaml,
-			Owner: "root",
-			Group: "root",
-		})
-	r.AddFile(
-		rpmpack.RPMFile{
-			Name:  "/usr/local/bin/velociraptor",
-			Body:  binary_content,
-			Mode:  0755,
-			Owner: "root",
-			Group: "root",
-		})
-	r.AddFile(
-		rpmpack.RPMFile{
-			Name:  "/etc/rc.d/init.d/velociraptor",
-			Body:  []byte(rpm_sysv_client_service_definition),
-			Mode:  0755,
-			Owner: "root",
-			Group: "root",
-		})
-
-	r.AddPrein(`
-getent group velociraptor >/dev/null || groupadd -g 115 -r velociraptor || :
-getent passwd velociraptor >/dev/null || \
-useradd -c "Privilege-separated Velociraptor" -u 115 -g velociraptor  -s /sbin/nologin \
-  -s /sbin/nologin -r -d /var/empty/velociraptor velociraptor 2> /dev/null || :
-`)
-
-	r.AddPostin("/sbin/chkconfig --add velociraptor")
-	r.AddPreun(`
-if [ "$1" = 0 ]
-then
-        /sbin/service velociraptor stop > /dev/null 2>&1 || :
-        /sbin/chkconfig --del velociraptor
-fi
-`)
-	r.AddPostun(`
-/sbin/service velociraptor start  > /dev/null 2>&1 || :
-`)
-
-	fd, err := os.OpenFile(output_path,
-		os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0600)
-	if err != nil {
-		return fmt.Errorf("Unable to create output file: %w", err)
-	}
-	defer fd.Close()
-
-	return r.Write(fd)
-}
-
 func init() {
 	command_handlers = append(command_handlers, func(command string) bool {
 		switch command {
 		case client_rpm_command.FullCommand():
-			if *client_rpm_command_use_sysv {
-				FatalIfError(client_rpm_command, doClientSysVRPM)
-			} else {
-				FatalIfError(client_rpm_command, doClientRPM)
-			}
+			FatalIfError(client_rpm_command, doClientRPM)
 
 		case server_rpm_command.FullCommand():
 			FatalIfError(server_rpm_command, doServerRPM)
@@ -633,4 +536,9 @@ func init() {
 		}
 		return true
 	})
+}
+
+func escape_sh(in string) string {
+	in = strings.ReplaceAll(in, "\\", "\\\\")
+	return strings.ReplaceAll(in, "$", "\\$")
 }
